Ensure no undef tags are set
This commit removes remaining usage of $env_tag. Without this change
undef tags are set causing a failure when comparing string with :undef
because puppet 4 cannot handle this case. Additionally, a correction is
made to in line documentation to reference the correctly titeld default.
Lastly, tag parameters are validated to ensure they are not undef. If
one is found to be undef then puppet fails reporting the parameter must
be set.