Update usage of zfs::share to zfs_share
This commit updates usage of the zfs::share define type to move to the zfs_share native type. As a part of this change the parameters are simplified in order to be compatible with the new type. Without this change zpr requires the usage of a depracated method of managing shares using zfs in Solaris.