Set tsp queue when value doesn't match
This commit sets the tsp queue via an exec when the queue size does not match what is configured. Without this change the .tsprc file is the only place this is set, which does not get checked upon first load of task spooler. The result is that after a reboot, or upon first boot, the queue size can end up smaller than configured because the default size is one.