Only mount volumes when running a job
This commit updates backup worker behavior to mount volumes at job
creation time. Without this change when a mount is created it is set to
ensure => mounted. The implication of this behavior is that when a
device backend is changed manual intervention is required to unmount the
currently used volume. This change supports placing the system into
maintenance mode, changing the backend storage configuration, and
letting the new volumes get mounted by the job as needed. The other
benefit of this change is it reduces load on the NFS server by
significantly reducing the number of simultaneous connections.